requirements.txt
transformers>=4.42.3
peft>=0.11.1
trl>=0.12.0
accelerate>=0.32.1
bitsandbytes>=0.45.2
datasets>=2.20.0
scikit-learn
pandas
wandb
rich
pyyaml